+++ src/index.js
@@ -1,3 +1,6 @@
-import {silkeyJwtGenerator} from "./SilkeyJwtGenerator.js";
-import {silkeyJwtPayloadVerificator} from "./SilkeyJwtPayloadVerificator.js";
-export { silkeyJwtGenerator, silkeyJwtPayloadVerificator };
+import * as Models from './models/index.js'
+import * as utils from './utils/index.js'
+
+import { generateSSORequestParams, tokenPayloadVerifier, messageToSign } from './sso.js'
+
+export default { Models, utils, generateSSORequestParams, tokenPayloadVerifier, messageToSign }