+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2020-11-04
+### Added
+- eslint
+- Methods for signing SSO request 
+- jsdocs
+- new model: `JwtPayload`
+- support for scope:email
+
+### Changed
+- Change package name from `silkey-jwt` to `silkey-sdk`
+
+### Removed 
+- Token generator
 
 ## [0.1.0] - 2020-10-06
 ### Added